gateway/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
kafka-python
fastavro
prometheus-client
